pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.3",
    "google-generativeai>=0.8.5",
    "numpy>=2.2.6",
    "openai>=2.8.1",
    "opencv-python>=4.12.0.88",
    "paddleocr>=3.3.2",
    "paddlepaddle>=3.2.2",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "pytesseract>=0.3.10",
    "reportlab>=4.4.5",
    "uvicorn>=0.38.0",
]
